feat: move sections tp separate pages, add router
