+++ modules/_fieldset.scss
@@ -17,7 +17,7 @@
 		margin-bottom: 0;
 	}
 
-	:where([type=checkbox i] + label, [type=radio i] + label){
+	:where([type='checkbox' i] + label, [type='radio' i] + label) {
 		display: inline;
 		margin: 0;
 	}
@@ -48,7 +48,6 @@
 %field-inline {
 	@include field-inline;
 }
-
 
 :where(legend) {
 	padding: 0 0.25em;
