+++ package.json
@@ -20,7 +20,7 @@
 		"test:e2e:ui": "playwright test --ui",
 		"lint": "eslint . --ext .vue,.js,.cjs,.mjs --fix --ignore-path .gitignore",
 		"format": "prettier --write modules/ *.cjs *.js *.scss",
-    "postversion": "git push && git push --tags"
+		"postversion": "git push && git push --tags"
 	},
 	"dependencies": {
 		"include-media": "^2.0.0"
@@ -38,7 +38,8 @@
 		"rimraf": "^5.0.1",
 		"sass": "^1.66.1",
 		"vite": "^4.4.9",
-		"vue": "^3.3.4"
+		"vue": "^3.3.4",
+		"vue-router": "^4.2.4"
 	},
 	"browserslist": [
 		"> 1%",
