+++ src/views/ExamplesButtons.vue
